+++ example/helloworld.py
@@ -6,9 +6,11 @@ from jobslib import BaseTask, argument, Config, option
 # settings --------------------------------------------------------------------
 
 CONFIG_CLASS = 'example.helloworld.HelloWorldConfig'
+
 ONE_INSTANCE = {
     'backend': 'jobslib.oneinstance.dummy.DummyLock',
-    }
+}
+
 
 # config ----------------------------------------------------------------------